requirements.txt
fastapi[standard]
httpx
psycopg2-binary
pydantic
pytest
python-dotenv
starlette
uvicorn
